preps for dockerizations
